feat: api 폴더 변경 및 일반상식 조회 api 세팅
