+++ src/api/scrap.ts
@@ -1,5 +1,6 @@
 import { WordScrapParams } from "@/types/scrapType";
 import { ScrapWordDTO } from "@/types/quizType";
+import { KnowledgeDTO, ScrapKnowledgeParams } from "@/types/commonType";
 
 import { instance } from "@/api/instance";
 
@@ -14,5 +15,24 @@ export const getScrapWord = async (userEmail: string) => {
   const { data } = await instance.get<ScrapWordDTO[]>(
     `/scrap/word-check?userEmail=${userEmail}`,
   );
+  return data;
+};
+
+export const scrapKnowledge = async ({
+  knowledgeId,
+  userEmail,
+}: ScrapKnowledgeParams) => {
+  const { data } = await instance.post(
+    `/scrap/knowledge-save?knowledgeId=${knowledgeId}&userEmail=${userEmail}`,
+  );
+
+  return data;
+};
+
+export const getScrapKnowledge = async (userEmail: string) => {
+  const { data } = await instance.get<KnowledgeDTO[]>(
+    `/scrap/knowledge-check?userEmail=${userEmail}`,
+  );
+
   return data;
 };
